Added StyleCop rules to Receipt. Fixed all errors and warnings
