--- src/Altinn.Platform/Altinn.Platform.Receipt/Altinn3.ruleset
@@ -1,177 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<RuleSet Name="New Rule Set" Description=" " ToolsVersion="15.0">
-  <Rules AnalyzerId="StyleCop.Analyzers" RuleNamespace="StyleCop.Analyzers">
-    <Rule Id="SA0001" Action="Error" />
-    <Rule Id="SA0002" Action="None" />
-    <Rule Id="SA1000" Action="Warning" />
-    <Rule Id="SA1001" Action="Warning" />
-    <Rule Id="SA1002" Action="Warning" />
-    <Rule Id="SA1003" Action="Warning" />
-    <Rule Id="SA1004" Action="Warning" />
-    <Rule Id="SA1005" Action="Warning" />
-    <Rule Id="SA1006" Action="Warning" />
-    <Rule Id="SA1007" Action="Warning" />
-    <Rule Id="SA1008" Action="Warning" />
-    <Rule Id="SA1009" Action="None" />
-    <Rule Id="SA1010" Action="Warning" />
-    <Rule Id="SA1011" Action="Warning" />
-    <Rule Id="SA1012" Action="Warning" />
-    <Rule Id="SA1013" Action="Warning" />
-    <Rule Id="SA1014" Action="Warning" />
-    <Rule Id="SA1015" Action="Warning" />
-    <Rule Id="SA1016" Action="Warning" />
-    <Rule Id="SA1017" Action="Warning" />
-    <Rule Id="SA1018" Action="Warning" />
-    <Rule Id="SA1019" Action="Warning" />
-    <Rule Id="SA1020" Action="Warning" />
-    <Rule Id="SA1021" Action="Warning" />
-    <Rule Id="SA1022" Action="Warning" />
-    <Rule Id="SA1023" Action="Warning" />
-    <Rule Id="SA1024" Action="None" />
-    <Rule Id="SA1025" Action="Warning" />
-    <Rule Id="SA1026" Action="Warning" />
-    <Rule Id="SA1027" Action="Warning" />
-    <Rule Id="SA1028" Action="None" />
-    <Rule Id="SA1100" Action="Error" />
-    <Rule Id="SA1101" Action="None" />
-    <Rule Id="SA1102" Action="Error" />
-    <Rule Id="SA1103" Action="Error" />
-    <Rule Id="SA1104" Action="Error" />
-    <Rule Id="SA1105" Action="Error" />
-    <Rule Id="SA1106" Action="Error" />
-    <Rule Id="SA1107" Action="Error" />
-    <Rule Id="SA1108" Action="Error" />
-    <Rule Id="SA1110" Action="Error" />
-    <Rule Id="SA1111" Action="Error" />
-    <Rule Id="SA1112" Action="Error" />
-    <Rule Id="SA1113" Action="Error" />
-    <Rule Id="SA1114" Action="Error" />
-    <Rule Id="SA1115" Action="Error" />
-    <Rule Id="SA1116" Action="Error" />
-    <Rule Id="SA1117" Action="Error" />
-    <Rule Id="SA1118" Action="Error" />
-    <Rule Id="SA1119" Action="Error" />
-    <Rule Id="SA1120" Action="Error" />
-    <Rule Id="SA1121" Action="Error" />
-    <Rule Id="SA1122" Action="Error" />
-    <Rule Id="SA1123" Action="None" />
-    <Rule Id="SA1124" Action="None" />
-    <Rule Id="SA1125" Action="Error" />
-    <Rule Id="SA1127" Action="Error" />
-    <Rule Id="SA1128" Action="None" />
-    <Rule Id="SA1129" Action="Error" />
-    <Rule Id="SA1130" Action="Error" />
-    <Rule Id="SA1131" Action="Error" />
-    <Rule Id="SA1132" Action="Error" />
-    <Rule Id="SA1133" Action="Error" />
-    <Rule Id="SA1134" Action="Error" />
-    <Rule Id="SA1135" Action="Error" />
-    <Rule Id="SA1136" Action="Error" />
-    <Rule Id="SA1137" Action="Warning" />
-    <Rule Id="SA1139" Action="None" />
-    <Rule Id="SA1200" Action="None" />
-    <Rule Id="SA1201" Action="None" />
-    <Rule Id="SA1202" Action="None" />
-    <Rule Id="SA1203" Action="None" />
-    <Rule Id="SA1204" Action="None" />
-    <Rule Id="SA1205" Action="None" />
-    <Rule Id="SA1206" Action="None" />
-    <Rule Id="SA1207" Action="None" />
-    <Rule Id="SA1208" Action="Error" />
-    <Rule Id="SA1209" Action="Error" />
-    <Rule Id="SA1210" Action="Error" />
-    <Rule Id="SA1211" Action="Error" />
-    <Rule Id="SA1212" Action="None" />
-    <Rule Id="SA1213" Action="None" />
-    <Rule Id="SA1214" Action="None" />
-    <Rule Id="SA1216" Action="Error" />
-    <Rule Id="SA1217" Action="Error" />
-    <Rule Id="SA1300" Action="Error" />
-    <Rule Id="SA1302" Action="Error" />
-    <Rule Id="SA1303" Action="Error" />
-    <Rule Id="SA1304" Action="Error" />
-    <Rule Id="SA1305" Action="Error" />
-    <Rule Id="SA1306" Action="Error" />
-    <Rule Id="SA1307" Action="Error" />
-    <Rule Id="SA1308" Action="Error" />
-    <Rule Id="SA1309" Action="None" />
-    <Rule Id="SA1310" Action="None" />
-    <Rule Id="SA1311" Action="Error" />
-    <Rule Id="SA1312" Action="Error" />
-    <Rule Id="SA1313" Action="None" />
-    <Rule Id="SA1314" Action="Error" />
-    <Rule Id="SA1400" Action="Warning" />
-    <Rule Id="SA1401" Action="Error" />
-    <Rule Id="SA1402" Action="None" />
-    <Rule Id="SA1403" Action="Error" />
-    <Rule Id="SA1404" Action="Error" />
-    <Rule Id="SA1405" Action="None" />
-    <Rule Id="SA1406" Action="None" />
-    <Rule Id="SA1407" Action="Error" />
-    <Rule Id="SA1408" Action="Error" />
-    <Rule Id="SA1410" Action="None" />
-    <Rule Id="SA1411" Action="None" />
-    <Rule Id="SA1413" Action="None" />
-    <Rule Id="SA1500" Action="Error" />
-    <Rule Id="SA1501" Action="Error" />
-    <Rule Id="SA1502" Action="Error" />
-    <Rule Id="SA1503" Action="Error" />
-    <Rule Id="SA1504" Action="Error" />
-    <Rule Id="SA1505" Action="Error" />
-    <Rule Id="SA1506" Action="Error" />
-    <Rule Id="SA1507" Action="Error" />
-    <Rule Id="SA1508" Action="Warning" />
-    <Rule Id="SA1509" Action="Warning" />
-    <Rule Id="SA1510" Action="Error" />
-    <Rule Id="SA1511" Action="Error" />
-    <Rule Id="SA1512" Action="Warning" />
-    <Rule Id="SA1513" Action="Error" />
-    <Rule Id="SA1514" Action="Error" />
-    <Rule Id="SA1515" Action="Error" />
-    <Rule Id="SA1516" Action="Error" />
-    <Rule Id="SA1517" Action="Error" />
-    <Rule Id="SA1518" Action="Error" />
-    <Rule Id="SA1519" Action="Error" />
-    <Rule Id="SA1520" Action="Error" />
-    <Rule Id="SA1600" Action="Warning" />
-    <Rule Id="SA1601" Action="Error" />
-    <Rule Id="SA1602" Action="Error" />
-    <Rule Id="SA1604" Action="Error" />
-    <Rule Id="SA1605" Action="Error" />
-    <Rule Id="SA1606" Action="Error" />
-    <Rule Id="SA1607" Action="Error" />
-    <Rule Id="SA1608" Action="Error" />
-    <Rule Id="SA1610" Action="Error" />
-    <Rule Id="SA1611" Action="None" />
-    <Rule Id="SA1612" Action="Error" />
-    <Rule Id="SA1613" Action="Error" />
-    <Rule Id="SA1614" Action="Error" />
-    <Rule Id="SA1615" Action="None" />
-    <Rule Id="SA1616" Action="None" />
-    <Rule Id="SA1617" Action="Error" />
-    <Rule Id="SA1618" Action="None" />
-    <Rule Id="SA1619" Action="None" />
-    <Rule Id="SA1620" Action="None" />
-    <Rule Id="SA1621" Action="Error" />
-    <Rule Id="SA1622" Action="None" />
-    <Rule Id="SA1623" Action="None" />
-    <Rule Id="SA1624" Action="Error" />
-    <Rule Id="SA1625" Action="Error" />
-    <Rule Id="SA1626" Action="Warning" />
-    <Rule Id="SA1627" Action="Error" />
-    <Rule Id="SA1629" Action="None" />
-    <Rule Id="SA1633" Action="None" />
-    <Rule Id="SA1634" Action="None" />
-    <Rule Id="SA1635" Action="None" />
-    <Rule Id="SA1636" Action="None" />
-    <Rule Id="SA1637" Action="None" />
-    <Rule Id="SA1638" Action="None" />
-    <Rule Id="SA1640" Action="None" />
-    <Rule Id="SA1641" Action="None" />
-    <Rule Id="SA1642" Action="None" />
-    <Rule Id="SA1643" Action="None" />
-    <Rule Id="SA1648" Action="Error" />
-    <Rule Id="SA1649" Action="Error" />
-    <Rule Id="SA1651" Action="Error" />
-  </Rules>
-</RuleSet>
