+++ src/Altinn.Platform/Altinn.Platform.Receipt/Tests/Testdata/Instances.cs
@@ -6,9 +6,8 @@ namespace Altinn.Platform.Receipt.Tests.Testdata
 {
     public static class Instances
     {
-        public static Instance Instance1 = new Instance
+        public static Instance Instance1 { get; set; } = new Instance
         {
-
             Id = "1000/1c3a4b9d-cbbe-4146-b370-4164e925812b",
             InstanceOwner = new InstanceOwner
             {
