+++ src/Altinn.Platform/Altinn.Platform.Receipt/Tests/Testdata/Parties.cs
@@ -4,7 +4,7 @@ namespace Altinn.Platform.Receipt.Tests.Testdata
 {
    public static class Parties
     {
-        public static Party Party1 = new Party
+        public static Party Party1 { get; set; } = new Party
         {
             PartyId = 50001,
             SSN = "12345678901",
