+++ src/Altinn.Platform/Altinn.Platform.Receipt/Tests/Testdata/UserProfiles.cs
@@ -4,8 +4,7 @@ namespace Altinn.Platform.Receipt.Tests.Testdata
 {
     public static class UserProfiles
     {
-
-        public static UserProfile User1 = new UserProfile
+        public static UserProfile User1 { get; set; } = new UserProfile
         {
             UserId = 1,
             Email = "test@test.no",
